move linkToProps
