+++ vue/components/SpaLink.vue
@@ -20,7 +20,7 @@
  * If you just want a simple piece-of-text link, though, this is your component.
  */
 
-import { type Link, linkToProps } from "@saflib/links";
+import { type Link, linkToProps } from "@saflib/vue";
 
 const { link } = defineProps<{ link: Link }>();
 const linkProps = linkToProps(link);
